Adding to output handlers.
This adds the log method for writing messages, and the close method for
tearing down the handler. Also adds SequenceOutputHandler for combining
multiple handlers (and removes RsyncCommand's ability to stack
handlers).

Also a few general changes: remove LOG_NAME option and add an error for
arguments that make no sense.